[DF-10017] Fixed compilation errors, win10x86
Former-commit-id: 0d1f79ffc6a9ffc1a52196b19648f063d7e001c1

+++ Projects/UnitTests/Classes/Tests/TypeReflectionTest.cpp
@@ -214,7 +214,7 @@ type_printer* get_type_printer(const DAVA::Type* type)
 
     static std::vector<type_printer> printers = {
         { DAVA::Type::Instance<int>(), [](char* buf, size_t bufsz, const DAVA::Any& any) { Snprintf(buf, bufsz, "%d", any.Get<int>()); } },
-        { DAVA::Type::Instance<size_t>(), [](char* buf, size_t bufsz, const DAVA::Any& any) { Snprintf(buf, bufsz, "%llu", any.Get<size_t>()); } },
+        { DAVA::Type::Instance<size_t>(), [](char* buf, size_t bufsz, const DAVA::Any& any) { Snprintf(buf, bufsz, "%llu", static_cast<uint64_t>(any.Get<size_t>())); } },
         { DAVA::Type::Instance<std::string>(), [](char* buf, size_t bufsz, const DAVA::Any& any) { Snprintf(buf, bufsz, "%s", any.Get<std::string>().c_str()); } }
     };
